ROBBIE!!
Wanted to let you know how the work was divided between the four of us:
Ria did the scaling, Zage created the stars, Jordan created the union and the stripes (red & white), and Georgia organized the flag and assigned the colors. We then worked together to combine our code. We each understand the code deeply, as you can tell by our writeup on Schoology, so hopefully this helps clear things up!